Check if current user

+++ app/controllers/api/comments_controller.rb
@@ -22,10 +22,12 @@ class Api::CommentsController < ApplicationController
 
   def destroy
     @comment = Comment.find(params[:id])
-    if @comment.destroy
-      render 'api/comments/show'
-    else
-      render json: @like.errors.full_messages, status: 422
+    if @comment.image.user == current_user
+      if @comment.destroy
+        render 'api/comments/show'
+      else
+        render json: @like.errors.full_messages, status: 422
+      end
     end
   end
